requirements.txt
python-dotenv
anthropic
langchain
langgraph

pymongo

# Jupyter Notebook
jupyter
ipykernel
ipywidgets
pyfiglet

#CLI interface
click
rich